reorg language selector
