+++ packages/ui/src/components/language-selector/index.ts
@@ -1,0 +1,2 @@
+export * from './language-dialog'
+export * from './types'
