+++ packages/ui/src/components/language-selector/language-dialog.tsx
@@ -2,57 +2,7 @@ import React, { FC, useEffect, useState } from 'react'
 
 import { Button, ButtonGroup, Dialog, Icon } from '@/components'
 
-export enum Language {
-  English = 'English',
-  Bulgarian = 'Bulgarian',
-  Croatian = 'Croatian',
-  Czech = 'Czech',
-  French = 'French',
-  German = 'German – Standard',
-  Irish = 'Irish – Extended',
-  Russian = 'Russian – QWERTY',
-  LatinAmerican = 'Latin American'
-}
-
-export enum LanguageCode {
-  EN = 'EN',
-  БГ = 'БГ',
-  HR = 'HR',
-  CZ = 'CZ',
-  FR = 'FR',
-  DE = 'DE',
-  IE = 'IE',
-  PY = 'PY',
-  LA = 'LA'
-}
-
-export interface LanguageInterface {
-  code: LanguageCode
-  name: Language
-}
-
-const languages: LanguageInterface[] = [
-  { code: LanguageCode.EN, name: Language.English },
-  { code: LanguageCode.БГ, name: Language.Bulgarian },
-  { code: LanguageCode.HR, name: Language.Croatian },
-  { code: LanguageCode.CZ, name: Language.Czech },
-  { code: LanguageCode.FR, name: Language.French },
-  { code: LanguageCode.DE, name: Language.German },
-  { code: LanguageCode.IE, name: Language.Irish },
-  { code: LanguageCode.PY, name: Language.Russian },
-  { code: LanguageCode.LA, name: Language.LatinAmerican }
-]
-
-interface LanguageDialogProps {
-  defaultLanguage?: LanguageCode
-  language?: LanguageCode
-  open: boolean
-  onOpenChange: (open: boolean) => void
-  onChange: (language: LanguageInterface) => void
-  onSave: (language: LanguageInterface) => void
-  onCancel: () => void
-  children: React.ReactNode
-}
+import { LanguageCode, LanguageDialogProps, languages } from './types'
 
 const LanguageDialog: FC<LanguageDialogProps> = ({
   defaultLanguage,
